requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
boto3==1.34.0
redis==5.0.1
asyncio-mqtt==0.16.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
aiofiles==23.2.1
pyyaml==6.0.1